+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 v2.2.0
 ------
 * Added a feature to stitch a local stack trace when reading throwables.
+* Added a feature to pass an uncaught exception directly to the handler.
 
 v2.1.0 (2023-01-06)
 ------